xds: remove MeshCaCertificateProvider and DynamicReloadingCertificate{Provider
