+++ xds/src/main/java/io/grpc/xds/internal/certprovider/CertificateProviderRegistry.java
@@ -40,8 +40,6 @@ public final class CertificateProviderRegistry {
       instance = new CertificateProviderRegistry();
       // TODO(sanjaypujare): replace with Java's SPI mechanism and META-INF resource
       instance.register(new FileWatcherCertificateProviderProvider());
-      instance.register(new DynamicReloadingCertificateProviderProvider());
-      instance.register(new MeshCaCertificateProviderProvider());
     }
     return instance;
   }
